// LDC.Atlas.WebApp/Client/app/shared/entities/directory-user.entity.ts
export class DirectoryUser {
    userId: string;
    firstName: string;
    lastName: string;
    emailAddress: string;
    phoneNumber: string;
    preferredLanguage: string;
    location: string;
    displayName: string;
    userPrincipalName: string;
    samAccountName: string;
    managerSamAccountName: string;
    companyRole: string;
}